backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.52
pydantic==1.10.7
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.6
email-validator
